run on a raspberry pi (cont)

+++ ui.py
@@ -11,7 +11,7 @@ SHUTDOWN = 0
 PLAY = 1
 PAUSE = 2
 GO_ON = 3
-STOP = 4
+NEXT = 4
 WAY_POINT = 5
 NEW_TRACK = 6
 
@@ -20,15 +20,16 @@ class Buzzer:
     def __init__(self):
         self.pi = pigpio.pi()
 
-        self.buzzer = False
+        # to stop it at start up if on !
+        self.buzzer = True
 
-    def start_buzzer(self):
+    def start(self):
         if not self.buzzer:
             self.pi.set_PWM_dutycycle(GPIO_BUZZER, 128)
             self.pi.set_PWM_frequency(GPIO_BUZZER, 1000)
             self.buzzer = True
 
-    def stop_buzzer(self):
+    def stop(self):
         if self.buzzer:
             self.pi.set_PWM_dutycycle(GPIO_BUZZER, 0)
             self.pi.set_PWM_frequency(GPIO_BUZZER, 0)
@@ -39,40 +40,44 @@ class Buzzer:
 
 
 class Inputs:
-    def __init__(self):
+    def __init__(self, cycles):
         self.pi = pigpio.pi()
+        self.cycles = cycles
+        self.current_cycles = 0
 
         self.pi.set_mode(GPIO_IN1, pigpio.INPUT)
         self.pi.set_mode(GPIO_IN2, pigpio.INPUT)
         self.pi.set_mode(GPIO_IN3, pigpio.INPUT)
         self.pi.set_mode(GPIO_IN4, pigpio.INPUT)
 
-        self.shutdown = False
-
     def read(self):
-        i1 = self.pi.read(GPIO_IN1)
-        i2 = self.pi.read(GPIO_IN2)
-        i3 = self.pi.read(GPIO_IN3)
-        i4 = self.pi.read(GPIO_IN4)
-        sum = i1 + i2 + i3 + i4
-        if sum == 1:
-            if i1 == 1:
-                return PLAY
-            elif i2 == 1:
-                return PAUSE
-            elif i3 == 1:
-                return GO_ON
-            elif i4 == 1:
-                return STOP
-        elif sum == 2:
-            if i1 + i2 == 2:
-                return WAY_POINT
-            if i2 + i3 == 2:
-                return NEW_TRACK
-        elif sum == 4:
-            self.shutdown = True
-            os.system("sudo shutdown --halt --no-wall")
-            return SHUTDOWN
+        if self.current_cycles == 0:
+            i1 = self.pi.read(GPIO_IN1)
+            i2 = self.pi.read(GPIO_IN2)
+            i3 = self.pi.read(GPIO_IN3)
+            i4 = self.pi.read(GPIO_IN4)
+            sum = i1 + i2 + i3 + i4
+            if sum == 1:
+                if i1 == 1:
+                    return PLAY
+                elif i2 == 1:
+                    return PAUSE
+                elif i3 == 1:
+                    return GO_ON
+                elif i4 == 1:
+                    return NEXT
+            elif sum == 2:
+                if i1 + i2 == 2:
+                    return WAY_POINT
+                if i2 + i3 == 2:
+                    return NEW_TRACK
+            elif sum == 4:
+                os.system("sudo shutdown --halt --no-wall")
+                return SHUTDOWN
+            if sum != 0:
+                self.current_cycles = self.cycles
+        else:
+            self.current_cycles -= 1
 
     def close(self):
         self.pi.stop()
